Consolidate v2 report APIs into one service
