+++ central/main.go
@@ -106,8 +106,7 @@ import (
 	processListeningOnPorts "github.com/stackrox/rox/central/processlisteningonport/service"
 	"github.com/stackrox/rox/central/pruning"
 	rbacService "github.com/stackrox/rox/central/rbac/service"
-	reportConfigurationService "github.com/stackrox/rox/central/reportconfigurations/service"
-	reportConfigurationServiceV2 "github.com/stackrox/rox/central/reportconfigurations/service/v2"
+	reportConfigurationService "github.com/stackrox/rox/central/reports/config/service"
 	vulnReportScheduleManager "github.com/stackrox/rox/central/reports/manager"
 	vulnReportV2Scheduler "github.com/stackrox/rox/central/reports/scheduler/v2"
 	reportService "github.com/stackrox/rox/central/reports/service"
@@ -408,7 +407,7 @@ func servicesToRegister() []pkgGRPC.APIService {
 
 	if env.VulnReportingEnhancements.BooleanSetting() {
 		// TODO Remove (deprecated) v1 report configuration service when Reporting enhancements are enabled by default.
-		servicesToRegister = append(servicesToRegister, reportConfigurationServiceV2.Singleton(), reportServiceV2.Singleton())
+		servicesToRegister = append(servicesToRegister, reportServiceV2.Singleton())
 	}
 
 	autoTriggerUpgrades := sensorUpgradeService.Singleton().AutoUpgradeSetting()
