+++ central/pruning/pruning_test.go
@@ -43,9 +43,9 @@ import (
 	roleMocks "github.com/stackrox/rox/central/rbac/k8srole/datastore/mocks"
 	k8sRoleBindingDataStore "github.com/stackrox/rox/central/rbac/k8srolebinding/datastore"
 	roleBindingMocks "github.com/stackrox/rox/central/rbac/k8srolebinding/datastore/mocks"
-	reportConfig "github.com/stackrox/rox/central/reportconfigurations/datastore"
-	reportConfigSearch "github.com/stackrox/rox/central/reportconfigurations/search"
-	reportConfigPgStore "github.com/stackrox/rox/central/reportconfigurations/store/postgres"
+	reportConfig "github.com/stackrox/rox/central/reports/config/datastore"
+	reportConfigSearch "github.com/stackrox/rox/central/reports/config/search"
+	reportConfigPgStore "github.com/stackrox/rox/central/reports/config/store/postgres"
 	reportDatastore "github.com/stackrox/rox/central/reports/snapshot/datastore"
 	reportSearch "github.com/stackrox/rox/central/reports/snapshot/datastore/search"
 	pgStore "github.com/stackrox/rox/central/reports/snapshot/datastore/store/postgres"
