+++ central/reports/config/datastore/datastore_impl.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/pkg/errors"
-	"github.com/stackrox/rox/central/reportconfigurations/search"
-	"github.com/stackrox/rox/central/reportconfigurations/store"
+	"github.com/stackrox/rox/central/reports/config/search"
+	"github.com/stackrox/rox/central/reports/config/store"
 	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/logging"
