+++ central/reports/config/service/service_impl.go
@@ -6,8 +6,8 @@ import (
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
 	"github.com/pkg/errors"
 	notifierDS "github.com/stackrox/rox/central/notifier/datastore"
-	"github.com/stackrox/rox/central/reportconfigurations/datastore"
-	"github.com/stackrox/rox/central/reportconfigurations/service/common"
+	"github.com/stackrox/rox/central/reports/config/datastore"
+	"github.com/stackrox/rox/central/reports/config/service/common"
 	"github.com/stackrox/rox/central/reports/manager"
 	collectionDS "github.com/stackrox/rox/central/resourcecollection/datastore"
 	v1 "github.com/stackrox/rox/generated/api/v1"
