+++ central/reports/snapshot/datastore/datastore_impl_test.go
@@ -6,7 +6,7 @@ import (
 	"context"
 	"testing"
 
-	reportConfigDS "github.com/stackrox/rox/central/reportconfigurations/datastore"
+	reportConfigDS "github.com/stackrox/rox/central/reports/config/datastore"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/env"
 	"github.com/stackrox/rox/pkg/fixtures"
